+++ src/main/java/cc/altius/FASP/model/rowMapper/ConsumptionBatchInfoRowMapper.java
@@ -26,7 +26,7 @@ public class ConsumptionBatchInfoRowMapper implements RowMapper<ConsumptionBatch
         }
         cb.setBatch(new Batch(rs.getInt("BATCH_ID"), rs.getInt("BATCH_PLANNING_UNIT_ID"), rs.getString("BATCH_NO"), rs.getBoolean("AUTO_GENERATED"), rs.getDate("EXPIRY_DATE")));
         cb.getBatch().setCreatedDate(rs.getDate("BATCH_CREATED_DATE"));
-        cb.setConsumptionQty(rs.getInt("BATCH_QTY"));
+        cb.setConsumptionQty(rs.getDouble("BATCH_QTY"));
         return cb;
     }
