+++ src/main/java/cc/altius/FASP/model/rowMapper/InventoryBatchInfoRowMapper.java
@@ -26,11 +26,11 @@ public class InventoryBatchInfoRowMapper implements RowMapper<InventoryBatchInfo
         }
         ib.setBatch(new Batch(rs.getInt("BATCH_ID"), rs.getInt("BATCH_PLANNING_UNIT_ID"), rs.getString("BATCH_NO"), rs.getBoolean("AUTO_GENERATED"), rs.getDate("EXPIRY_DATE")));
         ib.getBatch().setCreatedDate(rs.getDate("BATCH_CREATED_DATE"));
-        ib.setAdjustmentQty(rs.getInt("BATCH_ADJUSTMENT_QTY"));
+        ib.setAdjustmentQty(rs.getDouble("BATCH_ADJUSTMENT_QTY"));
         if (rs.wasNull()) {
             ib.setAdjustmentQty(null);
         }
-        ib.setActualQty(rs.getInt("BATCH_ACTUAL_QTY"));
+        ib.setActualQty(rs.getDouble("BATCH_ACTUAL_QTY"));
         if (rs.wasNull()) {
             ib.setActualQty(null);
         }
